refactor(ui): modularize localization catalog
