+++ ui/litellm-dashboard/src/i18n/locales/en/auth.ts
@@ -1,0 +1,17 @@
+export const enAuth = {
+  login: {
+    title: "Login",
+    username: "Username",
+    password: "Password",
+    submit: "Login",
+    submitting: "Logging in...",
+  },
+  onboarding: {
+    title: "Welcome to LiteLLM",
+    description: "Set up your account to continue.",
+  },
+  public: {
+    oauthCompleting: "Completing authorization...",
+    oauthComplete: "Authorization complete",
+  },
+} as const;
