+++ ui/litellm-dashboard/src/i18n/locales/en/common.ts
@@ -1,0 +1,22 @@
+export const enCommon = {
+  language: {
+    selectorLabel: "Language",
+    english: "English",
+    russian: "Русский",
+  },
+  actions: {
+    back: "Back",
+    cancel: "Cancel",
+    close: "Close",
+    continue: "Continue",
+    retry: "Retry",
+    save: "Save",
+    search: "Search",
+    submit: "Submit",
+  },
+  states: {
+    empty: "Nothing to show",
+    error: "Something went wrong",
+    loading: "Loading...",
+  },
+} as const;
