+++ ui/litellm-dashboard/src/i18n/locales/en/navigation.ts
@@ -1,0 +1,92 @@
+export const enNavigation = {
+  header: {
+    docs: "Docs",
+    chat: "Chat",
+  },
+  breadcrumbs: {
+    home: "Home",
+  },
+  sidebar: {
+    groups: {
+      "AI GATEWAY": "AI GATEWAY",
+      OBSERVABILITY: "OBSERVABILITY",
+      "ACCESS CONTROL": "ACCESS CONTROL",
+      "DEVELOPER TOOLS": "DEVELOPER TOOLS",
+      SETTINGS: "SETTINGS",
+    },
+    items: {
+      "api-keys": "Virtual Keys",
+      "llm-playground": "Playground",
+      models: "Models + Endpoints",
+      agentic: "Agentic",
+      agents: "Agents",
+      workflows: "Workflow Runs",
+      memory: "Memory",
+      "mcp-servers": "MCP Servers",
+      skills: "Skills",
+      guardrails: "Guardrails",
+      policies: "Policies",
+      tools: "Tools",
+      "search-tools": "Search Tools",
+      "vector-stores": "Vector Stores",
+      "tool-policies": "Tool Policies",
+      new_usage: "Usage",
+      "cost-optimization": "Cost Optimization",
+      logs: "Logs",
+      "guardrails-monitor": "Guardrails Monitor",
+      teams: "Teams",
+      projects: "Projects",
+      users: "Internal Users",
+      organizations: "Organizations",
+      "access-groups": "Access Groups",
+      budgets: "Budgets",
+      api_ref: "API Reference",
+      "model-hub-table": "AI Hub",
+      "learning-resources": "Learning Resources",
+      caching: "Response Cache",
+      experimental: "Experimental",
+      prompts: "Prompts",
+      "transform-request": "API Playground",
+      "tag-management": "Tag Management",
+      "4": "Old Usage",
+      settings: "Settings",
+      "router-settings": "Router Settings",
+      "logging-and-alerts": "Logging & Alerts",
+      "admin-panel": "Admin Settings",
+      "cost-tracking": "Cost Tracking",
+      "ui-theme": "UI Theme",
+    },
+    badges: { beta: "Beta", new: "New" },
+    controls: {
+      home: "LiteLLM home",
+      expand: "Expand sidebar",
+      collapse: "Collapse sidebar",
+    },
+    account: {
+      account: "Account",
+      accountMenu: "Account menu",
+      unknownRole: "Unknown role",
+      signedInAs: "signed in as",
+      unknownUser: "unknown",
+      tier: "Tier",
+      premium: "Premium",
+      standard: "Standard",
+      upgrade: "Upgrade to Premium for advanced features",
+      role: "Role",
+      email: "Email",
+      userId: "User ID",
+      copyEmail: "Copy email",
+      copyUserId: "Copy user ID",
+      thanks: "Thanks for using LiteLLM!",
+      hideNew: "Hide New Feature Indicators",
+      hideNewAria: "Toggle hide new feature indicators",
+      hidePrompts: "Hide All Prompts",
+      hidePromptsAria: "Toggle hide all prompts",
+      hideBlog: "Hide Blog Posts",
+      hideBlogAria: "Toggle hide blog posts",
+      hideIcon: "Hide Bouncing Icon",
+      hideIconAria: "Toggle hide bouncing icon",
+      logout: "Logout",
+    },
+  },
+} as const;
