+++ ui/litellm-dashboard/src/i18n/locales/ru/auth.ts
@@ -1,0 +1,17 @@
+export const ruAuth = {
+  login: {
+    title: "Вход",
+    username: "Имя пользователя",
+    password: "Пароль",
+    submit: "Войти",
+    submitting: "Выполняется вход...",
+  },
+  onboarding: {
+    title: "Добро пожаловать в LiteLLM",
+    description: "Настройте аккаунт, чтобы продолжить.",
+  },
+  public: {
+    oauthCompleting: "Завершаем авторизацию...",
+    oauthComplete: "Авторизация завершена",
+  },
+} as const;
