+++ ui/litellm-dashboard/src/i18n/locales/ru/common.ts
@@ -1,0 +1,22 @@
+export const ruCommon = {
+  language: {
+    selectorLabel: "Язык",
+    english: "English",
+    russian: "Русский",
+  },
+  actions: {
+    back: "Назад",
+    cancel: "Отмена",
+    close: "Закрыть",
+    continue: "Продолжить",
+    retry: "Повторить",
+    save: "Сохранить",
+    search: "Поиск",
+    submit: "Отправить",
+  },
+  states: {
+    empty: "Здесь пока ничего нет",
+    error: "Что-то пошло не так",
+    loading: "Загрузка...",
+  },
+} as const;
