+++ ui/litellm-dashboard/src/i18n/locales/ru/navigation.ts
@@ -1,0 +1,92 @@
+export const ruNavigation = {
+  header: {
+    docs: "Документация",
+    chat: "Чат",
+  },
+  breadcrumbs: {
+    home: "Главная",
+  },
+  sidebar: {
+    groups: {
+      "AI GATEWAY": "AI-ШЛЮЗ",
+      OBSERVABILITY: "НАБЛЮДАЕМОСТЬ",
+      "ACCESS CONTROL": "УПРАВЛЕНИЕ ДОСТУПОМ",
+      "DEVELOPER TOOLS": "ИНСТРУМЕНТЫ РАЗРАБОТЧИКА",
+      SETTINGS: "НАСТРОЙКИ",
+    },
+    items: {
+      "api-keys": "Виртуальные ключи",
+      "llm-playground": "Песочница",
+      models: "Модели и эндпоинты",
+      agentic: "Агентные функции",
+      agents: "Агенты",
+      workflows: "Запуски процессов",
+      memory: "Память",
+      "mcp-servers": "MCP-серверы",
+      skills: "Навыки",
+      guardrails: "Ограничители",
+      policies: "Политики",
+      tools: "Инструменты",
+      "search-tools": "Поиск инструментов",
+      "vector-stores": "Векторные хранилища",
+      "tool-policies": "Политики инструментов",
+      new_usage: "Использование",
+      "cost-optimization": "Оптимизация затрат",
+      logs: "Логи",
+      "guardrails-monitor": "Мониторинг ограничителей",
+      teams: "Команды",
+      projects: "Проекты",
+      users: "Внутренние пользователи",
+      organizations: "Организации",
+      "access-groups": "Группы доступа",
+      budgets: "Бюджеты",
+      api_ref: "Справочник API",
+      "model-hub-table": "Каталог AI",
+      "learning-resources": "Учебные материалы",
+      caching: "Кэш ответов",
+      experimental: "Экспериментальные",
+      prompts: "Промпты",
+      "transform-request": "API-песочница",
+      "tag-management": "Управление тегами",
+      "4": "Старое использование",
+      settings: "Настройки",
+      "router-settings": "Настройки роутера",
+      "logging-and-alerts": "Логи и оповещения",
+      "admin-panel": "Настройки администратора",
+      "cost-tracking": "Учёт затрат",
+      "ui-theme": "Тема интерфейса",
+    },
+    badges: { beta: "Бета", new: "Новое" },
+    controls: {
+      home: "Главная LiteLLM",
+      expand: "Развернуть сайдбар",
+      collapse: "Свернуть сайдбар",
+    },
+    account: {
+      account: "Аккаунт",
+      accountMenu: "Меню аккаунта",
+      unknownRole: "Неизвестная роль",
+      signedInAs: "выполнен вход как",
+      unknownUser: "неизвестный пользователь",
+      tier: "Уровень",
+      premium: "Премиум",
+      standard: "Стандартный",
+      upgrade: "Перейдите на Премиум для доступа к расширенным функциям",
+      role: "Роль",
+      email: "Эл. почта",
+      userId: "ID пользователя",
+      copyEmail: "Копировать эл. почту",
+      copyUserId: "Копировать ID пользователя",
+      thanks: "Спасибо, что используете LiteLLM!",
+      hideNew: "Скрыть метки новых функций",
+      hideNewAria: "Переключить отображение меток новых функций",
+      hidePrompts: "Скрыть все промпты",
+      hidePromptsAria: "Переключить отображение промптов",
+      hideBlog: "Скрыть записи блога",
+      hideBlogAria: "Переключить отображение записей блога",
+      hideIcon: "Скрыть прыгающий значок",
+      hideIconAria: "Переключить отображение прыгающего значка",
+      logout: "Выйти",
+    },
+  },
+} as const;
