+++ ui/litellm-dashboard/src/i18n/resources.ts
@@ -1,192 +1,1 @@
-export const resources = {
-  en: {
-    translation: {
-      language: {
-        selectorLabel: "Language",
-        english: "English",
-        russian: "Русский",
-      },
-      sidebar: {
-        groups: {
-          "AI GATEWAY": "AI GATEWAY",
-          OBSERVABILITY: "OBSERVABILITY",
-          "ACCESS CONTROL": "ACCESS CONTROL",
-          "DEVELOPER TOOLS": "DEVELOPER TOOLS",
-          SETTINGS: "SETTINGS",
-        },
-        items: {
-          "api-keys": "Virtual Keys",
-          "llm-playground": "Playground",
-          models: "Models + Endpoints",
-          agentic: "Agentic",
-          agents: "Agents",
-          workflows: "Workflow Runs",
-          memory: "Memory",
-          "mcp-servers": "MCP Servers",
-          skills: "Skills",
-          guardrails: "Guardrails",
-          policies: "Policies",
-          tools: "Tools",
-          "search-tools": "Search Tools",
-          "vector-stores": "Vector Stores",
-          "tool-policies": "Tool Policies",
-          new_usage: "Usage",
-          "cost-optimization": "Cost Optimization",
-          logs: "Logs",
-          "guardrails-monitor": "Guardrails Monitor",
-          teams: "Teams",
-          projects: "Projects",
-          users: "Internal Users",
-          organizations: "Organizations",
-          "access-groups": "Access Groups",
-          budgets: "Budgets",
-          api_ref: "API Reference",
-          "model-hub-table": "AI Hub",
-          "learning-resources": "Learning Resources",
-          caching: "Response Cache",
-          experimental: "Experimental",
-          prompts: "Prompts",
-          "transform-request": "API Playground",
-          "tag-management": "Tag Management",
-          "4": "Old Usage",
-          settings: "Settings",
-          "router-settings": "Router Settings",
-          "logging-and-alerts": "Logging & Alerts",
-          "admin-panel": "Admin Settings",
-          "cost-tracking": "Cost Tracking",
-          "ui-theme": "UI Theme",
-        },
-        badges: {
-          beta: "Beta",
-          new: "New",
-        },
-        controls: {
-          home: "LiteLLM home",
-          expand: "Expand sidebar",
-          collapse: "Collapse sidebar",
-        },
-        account: {
-          account: "Account",
-          accountMenu: "Account menu",
-          unknownRole: "Unknown role",
-          signedInAs: "signed in as",
-          unknownUser: "unknown",
-          tier: "Tier",
-          premium: "Premium",
-          standard: "Standard",
-          upgrade: "Upgrade to Premium for advanced features",
-          role: "Role",
-          email: "Email",
-          userId: "User ID",
-          copyEmail: "Copy email",
-          copyUserId: "Copy user ID",
-          thanks: "Thanks for using LiteLLM!",
-          hideNew: "Hide New Feature Indicators",
-          hideNewAria: "Toggle hide new feature indicators",
-          hidePrompts: "Hide All Prompts",
-          hidePromptsAria: "Toggle hide all prompts",
-          hideBlog: "Hide Blog Posts",
-          hideBlogAria: "Toggle hide blog posts",
-          hideIcon: "Hide Bouncing Icon",
-          hideIconAria: "Toggle hide bouncing icon",
-          logout: "Logout",
-        },
-      },
-    },
-  },
-  ru: {
-    translation: {
-      language: {
-        selectorLabel: "Язык",
-        english: "English",
-        russian: "Русский",
-      },
-      sidebar: {
-        groups: {
-          "AI GATEWAY": "AI-ШЛЮЗ",
-          OBSERVABILITY: "НАБЛЮДАЕМОСТЬ",
-          "ACCESS CONTROL": "УПРАВЛЕНИЕ ДОСТУПОМ",
-          "DEVELOPER TOOLS": "ИНСТРУМЕНТЫ РАЗРАБОТЧИКА",
-          SETTINGS: "НАСТРОЙКИ",
-        },
-        items: {
-          "api-keys": "Виртуальные ключи",
-          "llm-playground": "Песочница",
-          models: "Модели и эндпоинты",
-          agentic: "Агентные функции",
-          agents: "Агенты",
-          workflows: "Запуски процессов",
-          memory: "Память",
-          "mcp-servers": "MCP-серверы",
-          skills: "Навыки",
-          guardrails: "Ограничители",
-          policies: "Политики",
-          tools: "Инструменты",
-          "search-tools": "Поиск инструментов",
-          "vector-stores": "Векторные хранилища",
-          "tool-policies": "Политики инструментов",
-          new_usage: "Использование",
-          "cost-optimization": "Оптимизация затрат",
-          logs: "Логи",
-          "guardrails-monitor": "Мониторинг ограничителей",
-          teams: "Команды",
-          projects: "Проекты",
-          users: "Внутренние пользователи",
-          organizations: "Организации",
-          "access-groups": "Группы доступа",
-          budgets: "Бюджеты",
-          api_ref: "Справочник API",
-          "model-hub-table": "Каталог AI",
-          "learning-resources": "Учебные материалы",
-          caching: "Кэш ответов",
-          experimental: "Экспериментальные",
-          prompts: "Промпты",
-          "transform-request": "API-песочница",
-          "tag-management": "Управление тегами",
-          "4": "Старое использование",
-          settings: "Настройки",
-          "router-settings": "Настройки роутера",
-          "logging-and-alerts": "Логи и оповещения",
-          "admin-panel": "Настройки администратора",
-          "cost-tracking": "Учёт затрат",
-          "ui-theme": "Тема интерфейса",
-        },
-        badges: {
-          beta: "Бета",
-          new: "Новое",
-        },
-        controls: {
-          home: "Главная LiteLLM",
-          expand: "Развернуть сайдбар",
-          collapse: "Свернуть сайдбар",
-        },
-        account: {
-          account: "Аккаунт",
-          accountMenu: "Меню аккаунта",
-          unknownRole: "Неизвестная роль",
-          signedInAs: "выполнен вход как",
-          unknownUser: "неизвестный пользователь",
-          tier: "Уровень",
-          premium: "Премиум",
-          standard: "Стандартный",
-          upgrade: "Перейдите на Премиум для доступа к расширенным функциям",
-          role: "Роль",
-          email: "Эл. почта",
-          userId: "ID пользователя",
-          copyEmail: "Копировать эл. почту",
-          copyUserId: "Копировать ID пользователя",
-          thanks: "Спасибо, что используете LiteLLM!",
-          hideNew: "Скрыть метки новых функций",
-          hideNewAria: "Переключить отображение меток новых функций",
-          hidePrompts: "Скрыть все промпты",
-          hidePromptsAria: "Переключить отображение промптов",
-          hideBlog: "Скрыть записи блога",
-          hideBlogAria: "Переключить отображение записей блога",
-          hideIcon: "Скрыть прыгающий значок",
-          hideIconAria: "Переключить отображение прыгающего значка",
-          logout: "Выйти",
-        },
-      },
-    },
-  },
-} as const;
+export { resources } from "./catalog";
