+++ ui/litellm-dashboard/src/i18n/sidebar.ts
@@ -1,4 +1,4 @@
 import { SupportedLanguage } from "./language";
 import { resources } from "./resources";
 
-export const getSidebarTranslations = (language: SupportedLanguage) => resources[language].translation.sidebar;
+export const getSidebarTranslations = (language: SupportedLanguage) => resources[language].navigation.sidebar;
